ParseError does not longer has a code

+++ Tests/ParserCombinatorTests/Helpers.swift
@@ -9,9 +9,9 @@ import Foundation
 @testable import ParserCombinator
 
 internal struct TestError: ParseError, Equatable {
-    let code: UInt64
+    let code: Int
     
-    init(_ code: UInt64) {
+    init(_ code: Int) {
         self.code = code
     }
